tutorials/07-client-server-llm/pyproject.toml
[project]
name = "06-complete-server-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.12.3",
    "nest-asyncio>=1.6.0",
    "openai>=1.99.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "toml>=0.10.2",
]
